Move test to more appropriate test class
